Item741: contrib MANIFEST listed large number of files not in repository, so pseudo-install barfed. Also, MANIFEST lists directories, which confused pseudo-install (coded around)
git-svn-id: http://develop.twiki.org/svn/twiki/branches/DEVELOP@7398 a00a5322-12db-0310-a70b-8735589c885e

+++ pseudo-install.pl
@@ -65,7 +65,8 @@ sub copy_in {
 
 sub just_link {
     my( $moduleDir, $dir, $file ) = @_;
-    File::Path::mkpath( $dir );
+    return if -e $file;
+    File::Path::mkpath( $dir ) unless( -e $dir || -l $dir );
     my $argh = `ln -s $moduleDir/$file $file`;
     die "$argh $@" if ( $argh || $@ );
     print "Lunk $file\n";
